backend/requirements.txt
Django==5.0.2
django-cors-headers==4.3.1
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
graphene-django==3.1.6
PyJWT==2.8.0
google-generativeai==0.3.2
pyyaml==6.0.1
ruff==0.11.2
mecab-python3==1.0.6
django-ratelimit==4.1.0
